test only splitters

+++ sktime/forecasting/model_selection/_split.py
@@ -704,7 +704,8 @@ class CutoffSplitter(BaseSplitter):
     >>> ts = np.arange(10)
     >>> splitter = CutoffSplitter(fh=[2, 4], cutoffs=np.array([3, 5]), window_length=3)
     >>> list(splitter.split(ts))
-    [(array([1, 2, 3]), array([5, 7])), (array([3, 4, 5]), array([7, 9]))]
+    [(array([1, 2, 3], dtype=np.int64), array([5, 7], dtype=np.int64)),
+    (array([3, 4, 5], dtype=np.int64), array([7, 9], dtype=np.int64))]
     """
 
     def __init__(
